+++ src/stcal/jump/jump.py
@@ -1,7 +1,9 @@
 import logging
 import multiprocessing
 import time
+import warnings
 
+import numpy
 import numpy as np
 
 from . import constants
@@ -13,8 +15,19 @@ try:
 
     ELLIPSE_PACKAGE = 'opencv-python'
 except (ImportError, ModuleNotFoundError):
-    ELLIPSE_PACKAGE_WARNING = '`opencv-python` must be installed (`pip install stcal[opencv]`) ' \
-                              'in order to use ellipses'
+    try:
+        import shapely.geometry
+        import skimage.draw
+        import skimage.measure
+
+        from .circle import Circle
+
+        ELLIPSE_PACKAGE = 'scikit-image'
+        ELLIPSE_PACKAGE_WARNING = '`opencv-python` not installed; ' \
+                                  'using `scikit-image` + `shapely` for ellipse construction'
+    except (ImportError, ModuleNotFoundError):
+        ELLIPSE_PACKAGE_WARNING = 'an image processing package (either `opencv-python` or `scikit-image` + `shapely`)' \
+                                  'must be installed in order to use ellipses'
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
@@ -402,6 +415,10 @@ def extend_snowballs(plane, snowballs, sat_flag, jump_flag, expansion=1.5):
         color = (0, 0, 4)
         if ELLIPSE_PACKAGE == 'opencv-python':
             image = cv.circle(image, center, extend_radius, color, -1)
+        elif ELLIPSE_PACKAGE == 'scikit-image':
+            warnings.warn(ELLIPSE_PACKAGE_WARNING)
+            disk = skimage.draw.disk(center, extend_radius)
+            image[disk] = color
         else:
             raise ModuleNotFoundError(ELLIPSE_PACKAGE_WARNING)
         jump_circle = image[:, :, 2]
@@ -437,6 +454,10 @@ def extend_ellipses(plane, ellipses, sat_flag, jump_flag, expansion=1.1):
         color = (0, 0, 4)
         if ELLIPSE_PACKAGE == 'opencv-python':
             image = cv.ellipse(image, center, axes, alpha, 0, 360, color, -1)
+        elif ELLIPSE_PACKAGE == 'scikit-image':
+            warnings.warn(ELLIPSE_PACKAGE_WARNING)
+            ellipse = skimage.draw.ellipse(*center, *axes, rotation=alpha)
+            image[ellipse] = color
         else:
             raise ModuleNotFoundError(ELLIPSE_PACKAGE_WARNING)
         jump_ellipse = image[:, :, 2]
@@ -454,6 +475,10 @@ def find_circles(dqplane, bitmask, min_area):
         contours, hierarchy = cv.findContours(pixels, cv.RETR_EXTERNAL, cv.CHAIN_APPROX_SIMPLE)
         bigcontours = [con for con in contours if cv.contourArea(con) >= min_area]
         circles = [cv.minEnclosingCircle(con) for con in bigcontours]
+    elif ELLIPSE_PACKAGE == 'scikit-image':
+        contours = [shapely.geometry.Polygon(con) for con in skimage.measure.find_contours(pixels)]
+        bigcontours = [con for con in contours if con.area > min_area]
+        circles = [Circle.from_points(numpy.stack(con.exterior.xy, axis=1)) for con in bigcontours]
     else:
         raise ModuleNotFoundError(ELLIPSE_PACKAGE_WARNING)
     return circles
@@ -463,14 +488,31 @@ def find_ellipses(dqplane, bitmask, min_area):
     # Using an input DQ plane this routine will find the groups of pixels with at least the minimum
     # area and return a list of the minimum enclosing ellipse parameters.
     pixels = np.bitwise_and(dqplane, bitmask)
+
     if ELLIPSE_PACKAGE == 'opencv-python':
         contours, hierarchy = cv.findContours(pixels, cv.RETR_EXTERNAL, cv.CHAIN_APPROX_SIMPLE)
         bigcontours = [con for con in contours if cv.contourArea(con) > min_area]
         # minAreaRect is used becuase fitEllipse requires 5 points and it is possible to have a contour
         # with just 4 points.
         ellipses = [cv.minAreaRect(con) for con in bigcontours]
+    elif ELLIPSE_PACKAGE == 'scikit-image':
+        contours = [shapely.geometry.Polygon(con) for con in skimage.measure.find_contours(pixels)]
+        bigcontours = [con for con in contours if con.area > min_area]
+        rectangles = [
+            numpy.flip(numpy.stack(con.minimum_rotated_rectangle.exterior.xy, axis=1), axis=1)
+            for con in bigcontours
+        ]
+        ellipses = [
+            (
+                numpy.mean(rectangle[[0, 2], :], axis=0),
+                numpy.hypot(*numpy.diff(rectangle[[0, 1, 2], :], axis=0)),
+                numpy.degrees(numpy.arctan2(*numpy.flip(numpy.diff(rectangle[[3, 0], :], axis=0)[0])))
+            )
+            for rectangle in rectangles
+        ]
     else:
         raise ModuleNotFoundError(ELLIPSE_PACKAGE_WARNING)
+
     return ellipses